Move the enumeration from inside a class to toplevel.
